+++ src/ctypes/cstubs_internals.mli
@@ -42,6 +42,7 @@ type 'a typ = 'a Ctypes_static.typ =
   | View            : ('a, 'b) view             -> 'a typ
   | Array           : 'a typ * int              -> 'a Ctypes_static.carray typ
   | Bigarray        : (_, 'a, _) Ctypes_bigarray.t -> 'a typ
+  | OCamlUnsafe     : 'a ocaml_type             -> 'a ocaml typ
   | OCaml           : 'a ocaml_type             -> 'a ocaml typ
 and ('a, 'b) pointer = ('a, 'b) Ctypes_static.pointer =
   CPointer : 'a typ Ctypes_ptr.Fat.t -> ('a, [`C]) pointer