+++ src/ctypes/ctypes_memory.ml
@@ -34,6 +34,7 @@ let rec build : type a b. a typ -> b typ Fat.t -> a
     | View { read; ty } ->
       let buildty = build ty in
       (fun buf -> read (buildty buf))
+    | OCamlUnsafe _
     | OCaml _ -> (fun buf -> assert false)
     (* The following cases should never happen; non-struct aggregate
        types are excluded during type construction. *)
@@ -72,6 +73,7 @@ let rec write : type a b. a typ -> a -> b Fat.t -> unit
     | View { write = w; ty } ->
       let writety = write ty in
       (fun v -> writety (w v))
+    | OCamlUnsafe _ -> raise IncompleteType
     | OCaml _ -> raise IncompleteType
 
 let null : unit ptr = CPointer (Fat.make ~reftyp:Void Raw.null)
@@ -90,6 +92,7 @@ let rec (!@) : type a. a ptr -> a
         { astart = CPointer (Fat.coerce cptr elemtype); alength }
       | Bigarray b -> Ctypes_bigarray.view b cptr
       | Abstract _ -> { structured = ptr }
+      | OCamlUnsafe _ -> raise IncompleteType
       | OCaml _ -> raise IncompleteType
       (* If it's a value type then we cons a new value. *)
       | _ -> build (Fat.reftype cptr) cptr